requirements.txt
flask
numpy
keras
pandas
opencv-python
pyttsx3
tensorflow
google.generativeai
SpeechRecognition